Add initial test, JSDoc configuration
This test alone gets over 90% coverage:

```text
----------|---------|----------|---------|---------
File      | % Stmts | % Branch | % Funcs | % Lines
----------|---------|----------|---------|---------
All files |   90.95 |    85.71 |    61.9 |   90.95
 index.js |   90.95 |    85.71 |    61.9 |   90.95
----------|---------|----------|---------|---------
```

It is technically a "large" test since it exercises the main UI and
touches almost everything in the process. Of course I'm going to
refactor and write more fine grained tests. But it is nice to have a
fast smoke test to keep me honest and focused as I start to do it.