Test collectPartials(), ignore @partial-block
While adding tests for PartialCollector, it became apparent that
wrapping it in collectPartials() made both the tests and the production
code more concise.

As expected, the code already handled block partials and ignored dynamic
and inline partials. But while reading the partials documentation, I
discovered that the code didn't yet handle `@partial-block`:

- https://handlebarsjs.com/guide/partials.html

It proved easy enough to test and fix.