Bump to v1.0.1, add docs, use tsc+tsconfig.json
Greatly expanded the README documentation and updated some of the JSDoc
comments. Had to adjust a couple of JSDoc @type casts as well with
regard to Handlebars.CompileOptions and Handlebars.PrecompileOptions.

Also switched from jsconfig.json to tsconfig.json, as it seems better
supported out of the box. Adjusted `pnpm typecheck` to reflect this as
well.

Bumped vitest to v1.2.0.