Bump to v1.0.2, improve template.d.ts, bump vitest
The previous version of lib/template.d.ts worked, and enabled Visual
Studio Code to provide documentation in JSDoc comments like:

```js
* @param {import("./component.hbs").TemplateRenderer} [params.render] -
```

or:

```js
* @param {import("*.hbs").TemplateRenderer} [params.render] -
```

But the 'jsdoc' tool, even with 'jsdoc-plugin-typescript' installed,
would choke on them. This update fixes this, and allows for a better
experience all around by providing a proper namespace:

```js
* @param {HandlebarsPrecompiler.TemplateRenderer} [params.render] -
```

I tried to think through how to declare the namespace so that users
wouldn't have to copy it. However, the original comment in the README
still stands:

> This is necessary because the precompiled modules are generated in
> _your_ project, not in `rollup-plugin-handlebars-precompiler`, so
> that's where TypeScript needs to find the type declarations.

Also bumped vitest to v1.2.1, along with:

- @stylistic/eslint-plugin-js: v1.5.4
- @types/node: v20.11.5